+++ gif/__init__.py
@@ -1,0 +1,19 @@
+from .gif import Gif
+from .frame import Frame
+from .comment_extension import CommentExtension
+from .plain_text_extension import PlainTextExtension
+from .application_extension import ApplicationExtension
+from .graphic_control_extension import GraphicControlExtension
+from .incorrect_file_format import IncorrectFileFormat
+from .disposal_method import DisposalMethod
+
+__all__ = [
+    "Gif",
+    "Frame",
+    "CommentExtension",
+    "PlainTextExtension",
+    "ApplicationExtension",
+    "IncorrectFileFormat",
+    "DisposalMethod",
+    "GraphicControlExtension"
+]
