+++ gif/application_extension.py
@@ -1,0 +1,8 @@
+from attrs import define
+
+
+@define
+class ApplicationExtension:
+    application_name: str | None = None
+    identify: str | None = None
+    data: bytes | None = None
