+++ gif/comment_extension.py
@@ -1,0 +1,6 @@
+from attrs import define, field
+
+
+@define
+class CommentExtension:
+    data: bytes = field(default=None)
