+++ gif/incorrect_file_format.py
@@ -1,0 +1,3 @@
+class IncorrectFileFormat(Exception):
+    def __init__(self, message):
+        super().__init__(message)
