+++ gif/plain_text_extension.py
@@ -1,0 +1,15 @@
+from attrs import define, field
+
+
+@define
+class PlainTextExtension:
+    top: int = field(default=None)
+    left: int = field(default=None)
+    width: int = field(default=None)
+    height: int = field(default=None)
+
+    char_width: int = field(default=None)
+    char_height: int = field(default=None)
+    background_color: int = field(default=None)
+    text_color: int = field(default=None)
+    data: bytes = field(default=None)
