+++ lzw/__init__.py
@@ -1,0 +1,7 @@
+from .decoder import lzw_decode
+from .encoder import lzw_encode
+
+__all__ = [
+    "lzw_decode",
+    "lzw_encode"
+]
