+++ parser/__init__.py
@@ -1,0 +1,7 @@
+from .writer import write_gif
+from .reader import decode_gif
+
+__all__ = [
+    "write_gif",
+    "decode_gif"
+]
